Fix missing quotes in the sentry conf and update repo URL
In the configmap for the Sentry config, there're some places that miss to use the `quote` function and would cause some problems in certain cases.
One example is when the admin email look like `#hello@example.com` (this is an valid email)
I also updated the repo URL for stable. See this announcement: https://helm.sh/blog/new-location-stable-incubator-charts/