Add support for multiple external kafkas
The main motivation for this change is to enable AWS MSK used as
external kafka. This change keeps config backward compatible, which
means if a simple structure given under the `externalKafka` then it used
as before. However if a list of these structures given, then all of the
items used.

Example inputs:
- simple:
  ```
  externalKafka:
    host: broker-1
    port: 9094
  ```

- multiple kafka brokers:
  ```yaml
  externalKafka:
    - host: broker-1
      port: 9094
    - host: broker-2
      port: 9094
    - host: broker-3
      port: 9094
  ```